Tutorial#(023 Displaying child components)
Skiday list and skiday roes were our child process.
The <tr> tag defines a row in an HTML table
The <td> tag defines a standard cell in an HTML table.
The <th> tag defines a header cell in an HTML table.
The <tbody> tag is used to group the body content in an HTML table.
Whenever we are rendering a childeren in aray we have to supply a key value
so that react can know which things to be update..